style(test): use Tuple2 field access for consistency with production code
Switch counts._1().longValue() / _2().longValue() to counts._1 / _2 in the
test, matching the field-access style used in the production code path
(StreamSync's errorTableCounts._1) and elsewhere in StreamSync.java (e.g.
safeCreateRDDs._1). Long auto-unboxes via the assertEquals(long, long)
overload. Addresses an inline nit from the AI reviewer.

+++ hudi-utilities/src/test/java/org/apache/hudi/utilities/streamer/TestStreamSyncWriteStatusValidation.java
@@ -48,8 +48,8 @@ class TestStreamSyncWriteStatusValidation extends SparkClientFunctionalTestHarne
 
     Tuple2<Long, Long> counts = StreamSync.sumRecordAndErrorCounts(writeStatusRDD);
 
-    assertEquals(37L, counts._1().longValue(), "total records summed across all partitions");
-    assertEquals(8L, counts._2().longValue(), "total errored records summed across all partitions");
+    assertEquals(37L, counts._1, "total records summed across all partitions");
+    assertEquals(8L, counts._2, "total errored records summed across all partitions");
   }
 
   @Test
@@ -58,8 +58,8 @@ class TestStreamSyncWriteStatusValidation extends SparkClientFunctionalTestHarne
 
     Tuple2<Long, Long> counts = StreamSync.sumRecordAndErrorCounts(emptyRDD);
 
-    assertEquals(0L, counts._1().longValue());
-    assertEquals(0L, counts._2().longValue());
+    assertEquals(0L, counts._1);
+    assertEquals(0L, counts._2);
   }
 
   @Test
@@ -70,8 +70,8 @@ class TestStreamSyncWriteStatusValidation extends SparkClientFunctionalTestHarne
 
     Tuple2<Long, Long> counts = StreamSync.sumRecordAndErrorCounts(emptyRDD);
 
-    assertEquals(0L, counts._1().longValue());
-    assertEquals(0L, counts._2().longValue());
+    assertEquals(0L, counts._1);
+    assertEquals(0L, counts._2);
   }
 
   private static WriteStatus writeStatus(long totalRecords, long totalErrorRecords) {
